test(commands): #1147 認可回帰を分離
